% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{addClipped}
\alias{addClipped}
\alias{addClipped,GAlignments-method}
\title{Extrapolates mapping location from clipped, aligned reads}
\usage{
addClipped(bam, ...)

\S4method{addClipped}{GAlignments}(bam, ...)
}
\arguments{
\item{bam}{A set of aligned reads}

\item{bam}{A GAlignments object}
}
\value{
A \code{\link[GenomicRanges]{GRanges}} representation of the extended
mapping locations
}
\description{
Extrapolates the mapping location of a read by assuming that
the clipped regions should map adjacent to the mapped locations.  This
is not always a good assumption!
}
\author{
Helen Lindsay
}

