% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAlignments.R
\name{addCodonFrame}
\alias{addCodonFrame}
\title{Internal CrispRVariants function for indicating codon frame on an
alignment tile plot}
\usage{
addCodonFrame(p, width, codon.frame)
}
\arguments{
\item{p}{A ggplot object, typically from CrispRVariants:::makeAlignmentTilePlot}

\item{width}{The number of nucleotides in the alignments}

\item{codon.frame}{The leftmost starting location of the next codon - 1,2,or 3}
}
\value{
A ggplot object with added vertical lines indicating the frame
}
\description{
Adds vertical dotted lines in intervals of three nucleotides.
Codon frame is supplied, alignments are assumed not to span an intron-exon
junction.
}
\author{
Helen Lindsay
}
