% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alleleLabels.R
\name{alleleLabels}
\alias{alleleLabels}
\title{alleleLabels}
\usage{
alleleLabels(alns, target, reference, indel.ranges = NULL,
  indel.freq = NULL, snp.ranges = NULL, snp.freq = NULL,
  genome.to.pos = NULL, novar.label = "No variant",
  mismatch.label = "SNV:", ...)
}
\arguments{
\item{alns}{(GenomicAlignments)}

\item{target}{(GRanges(1))}

\item{indel.ranges}{(IRanges)}

\item{indel.freq}{(numeric)}

\item{snp.ranges}{(IRanges)}

\item{genome.to.pos}{(numeric)}

\item{novar.label}{(Character(1)) Label for non-variant reads
(Default: No variant)}

\item{mismatch.label}{(Character(1)) Prefix for single nucleotide
variants (Default: SNV:)}

\item{...}{Extra arguments for individual counting functions}
}
\description{
Developmental function for more flexible labeling
of variant alleles
}
