% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\docType{methods}
\name{alleles}
\alias{alleles}
\alias{alleles,CrisprSet-method}
\title{Get allele names}
\usage{
alleles(obj, ...)

\S4method{alleles}{CrisprSet}(obj, ...)
}
\arguments{
\item{obj}{An object containing variant alleles}

\item{...}{additional arguments}
}
\value{
A data frame relating CIGAR strings to variant labels
}
\description{
Function to access allele names
}
\examples{
data("gol_clutch1")
alleles <- alleles(gol)
}
\author{
Helen Lindsay
}
