% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\docType{methods}
\name{alns}
\alias{alns}
\alias{alns,CrisprSet-method}
\title{Get alignments}
\usage{
alns(obj, ...)

\S4method{alns}{CrisprSet}(obj, ...)
}
\arguments{
\item{obj}{An object containing aligned sequences}

\item{...}{additional arguments}
}
\value{
A GAlignmentsList of consensus sequences on the positive strand.
}
\description{
Return alignments from an object that contains them.
For a CrisprSet object, these are truncated, non-chimeric alignments
}
\examples{
data("gol_clutch1")
alns <- alns(gol)
}
\author{
Helen Lindsay
}
