% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotVariants.R
\name{annotateGenePlot}
\alias{annotateGenePlot}
\title{Plots gene structure and annotates with target location}
\usage{
annotateGenePlot(txdb, target, target.colour = "red", target.size = 1,
  gene.text.size = 12, panel.margin = unit(c(0.1, 0.1, 0.1, 0.1), "lines"))
}
\arguments{
\item{txdb}{A GenomicFeatures:TxDb object}

\item{target}{Location of target (GRanges)}

\item{target.colour}{Colour of box indicating targt region}

\item{target.size}{Thickness of box indicating target region}

\item{gene.text.size}{Size for figure label}

\item{panel.margin}{Unit object, margin size}
}
\description{
Uses ggbio to plot the gene structure, annotates this with the
target location
}

