% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotVariants.R
\name{arrangePlots}
\alias{arrangePlots}
\title{Arrange plots for plotVariants:CrisprSet}
\usage{
arrangePlots(top.plot, left.plot, right.plot, fig.height = NULL,
  col.wdth.ratio = c(2, 1), row.ht.ratio = c(1, 6),
  left.plot.margin = unit(c(0.25, 0, 10, 0.5), "lines"))
}
\arguments{
\item{top.plot}{ggplot grob, placed on top of the figure, spanning the figure width}

\item{left.plot}{ggplot, placed in the second row on the left}

\item{right.plot}{ggplot, placed in the second row on the right.
y-axis labels are removed.}

\item{fig.height}{Actual height for the figure. If not provided,
figure height is the sum of the row.ht.ratio (Default: NULL)}

\item{col.wdth.ratio}{Vector specifying column width ratio (Default: c(2, 1))}

\item{row.ht.ratio}{Vector specifying row height ratio (Default: c(1,6))}

\item{left.plot.margin}{Unit object specifying margins of left.plot.
Margins of right.plot are constrained by the left.plot.}
}
\description{
Arranges 3 plots in two rows.  The vertical margins of the
left.plot and right.plot constrained to be equal
}

