% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\docType{methods}
\name{countDeletions}
\alias{countDeletions}
\alias{countDeletions,GAlignments-method}
\title{Count the number of reads containing a deletion}
\usage{
countDeletions(alns, ...)

\S4method{countDeletions}{GAlignments}(alns, multi.del = FALSE,
  del.and.ins = FALSE, del.ops = c("D"))
}
\arguments{
\item{alns}{The aligned reads}

\item{multi.del}{If TRUE, returns the exact number of deletions,
i.e., if one read contains 2 deletions, it contributes 2 to the
total count (default: FALSE)}

\item{del.and.ins}{If TRUE, counts deletions regardless of whether
reads also contain insertions.  If FALSE, counts reads that contain
deletions but not insertions (default: FALSE)}

\item{del.ops}{Cigar operations counted as deletions.  Default: c("D")}
}
\description{
Counts the number of reads containing a deletion of any size
in a set of aligned reads
}
\author{
Helen Lindsay
}

