% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{countIndels}
\alias{countIndels}
\title{Counts the number of reads containing an insertion and/or deletion}
\usage{
countIndels(alns)
}
\arguments{
\item{alns}{The alignments}
}
\description{
Counts the number of target reads that include at least one
insertion or deletion, by counting cigar operations "I" (insertion),
"D" (deletion) and "N" (junction operation used by some aligners)
}
\author{
Helen Lindsay
}

