% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\docType{methods}
\name{countInsertions}
\alias{countInsertions}
\alias{countInsertions,GAlignments-method}
\title{Counts the number of reads containing an insertion}
\usage{
countInsertions(alns, ...)

\S4method{countInsertions}{GAlignments}(alns, ins.and.del = FALSE,
  multi.ins = FALSE, del.ops = c("D"))
}
\arguments{
\item{alns}{The aligned reads}

\item{ins.and.del}{If TRUE, counts insertions regardless of whether
reads also contain deletions  If FALSE, counts reads that contain
insertions but not deletions (default: FALSE)}

\item{multi.ins}{If TRUE, returns the exact number of insertions,
i.e., if one read contains 2 insertions, it contributes 2 to the
total count (default: FALSE)}

\item{del.ops}{Cigar operations counted as deletions.  Default: c("D")}
}
\description{
Counts the number of reads containing an insertion of any size
in a set of aligned reads
}
\author{
Helen Lindsay
}

