% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CrisprRun.R
\docType{class}
\name{CrisprRun-class}
\alias{CrisprRun}
\alias{CrisprRun-class}
\title{CrisprRun class}
\arguments{
\item{bam}{a GAlignments object containing (narrowed) alignments to the target region.
Filtering of the bam should generally be done before initialising a CrisprRun object}

\item{target}{The target location, a GRanges object}

\item{genome.ranges}{A GRangesList of genomic coordinates for the cigar operations.
If bam is a standard GAlignments object, this is equivalent to
cigarRangesAlongReferenceSpace + start(bam)}

\item{rc}{(reverse complement)  Should the alignments be reverse complemented,
i.e. displayed with respect to the negative strand?  (Default: FALSE)}

\item{name}{A name for this set of reads, used in plots if present (Default: NULL)}

\item{chimeras}{Off-target chimeric alignments not in bam.  (Default: empty)}

\item{verbose}{Print information about initialisation progress (Default: TRUE)}
}
\description{
A ReferenceClass container for a single sample of alignments narrowed
to a target region
}
\section{Fields}{

\describe{
\item{\code{alns}}{A GAlignments object containing the narrowed reads.  Note that if the alignments
are represented with respect to the reverse strand, the "start" remains with repect to the
forward strand, whilst the cigar and the sequence are reverse complemented.}

\item{\code{name}}{The name of the sample}

\item{\code{cigar_labels}}{A vector of labels for the reads, based on the cigar strings,
optionally renumbered with respect to a new zero point (e.g. the cut site) and
shortened to only insertion and deletion locations.
Set at initialisation of a CrisprSet object, but not at
initialisation of a CrisprRun object.}

\item{\code{chimeras}}{Chimeric, off-target alignments corresponding to alignments in alns}
}}
\section{Methods}{

\describe{
\item{\code{removeSeqs(idxs)}}{Description:
  Remove sequences from a CrisprRun object and from the internal CrisprRun 
  fields that store insertion locations for plotting.

Input parameters:
  idxs:     Indexes of reads to remove}
}}
\author{
Helen Lindsay
}
\seealso{
\code{\link[crispRvariants]{CrisprSet}}
}

