% Generated by roxygen2 (4.0.2): do not edit by hand
\name{excludeFromBam}
\alias{excludeFromBam}
\title{Removes reads from a bam file}
\usage{
excludeFromBam(bam, exclude_ranges = GRanges(), exclude_names = NA)
}
\arguments{
\item{bam}{a GAlignments object}

\item{exclude_ranges}{}

\item{exclude_names}{}
}
\description{
Returns a GAlignments excluding reads based on either name and/or location
}
\author{
Helen Lindsay
}

