% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/excludeFromBam.R
\name{excludeFromBamByName}
\alias{excludeFromBamByName}
\title{Removes reads from a bam file}
\usage{
excludeFromBamByName(bam, exclude.names)
}
\arguments{
\item{bam}{A GAlignments object}

\item{exclude.names}{a character vector of sequence names to exclude}
}
\description{
Returns a GAlignments excluding reads based on name
}
\author{
Helen Lindsay
}

