% Generated by roxygen2 (4.0.2): do not edit by hand
\name{excludeFromBamByRange}
\alias{excludeFromBamByRange}
\title{Removes reads from a bam file}
\usage{
excludeFromBamByRange(bam, exclude_ranges)
}
\arguments{
\item{bam}{a GAlignments object}

\item{exclude_ranges}{A GRanges object containing regions to exclude}
}
\description{
Returns a GAlignments excluding reads based on location
}
\author{
Helen Lindsay
}

