% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.explodeCigarOpCombs}
\alias{.explodeCigarOpCombs}
\title{.explodeCigarOpCombs}
\usage{
.explodeCigarOpCombs(cigar, ops = GenomicAlignments::CIGAR_OPS)
}
\arguments{
\item{cigar}{character(m) A vector of cigar strings}

\item{ops}{character(n) Which operations should be kept?}
}
\value{
The operations, as a CharacterList
}
\description{
Breaks cigar strings into individual operations
}
\author{
Helen Lindsay
}
