% Generated by roxygen2 (4.0.2): do not edit by hand
\name{findChimeras}
\alias{findChimeras}
\title{Find chimeric reads}
\usage{
findChimeras(bam)
}
\arguments{
\item{bam}{A GAlignments object}
}
\description{
Assumes the the GAlignments object does not contain multimapping reads.
That is, read names that appear more than ones in the file are considered
chimeras.
}
\author{
Helen Lindsay
}

