% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/initialisers.R
\name{findDeletions}
\alias{findDeletions}
\title{Find location of deletions that overlap a target region.}
\usage{
findDeletions(target.start, target.end, alns, ref.ranges, cig.ops,
  del.chars = c("D", "N"))
}
\arguments{
\item{target.start}{The target start location}

\item{target.end}{The target end location}

\item{alns}{GAlignments, only containing alignments spanning the target region}

\item{ref.ranges}{IRangesList created with GenomicAlignments::cigarRangesAlongReferenceSpace}

\item{cig.ops}{Cigar operations, created with GenomicAlignments::explodeCigarOps}

\item{del.chars}{Characters that may represent deletions.  Default: c("D", "N")}
}
\value{
A list of the start and end locations w.r.t. the reads for passing to
GenomicAlignments::CigarNarrow
}
\description{
For reads with a deletion spanning one or both
ends of the target location, narrow the alignment to encompass the deletion
Deletions may be coded as either "D" or "N" (splice junction),
}

