% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\docType{methods}
\name{findSNVs}
\alias{findSNVs}
\alias{findSNVs,CrisprSet-method}
\title{Find frequent SNVs}
\usage{
findSNVs(obj, ...)

\S4method{findSNVs}{CrisprSet}(obj, ..., freq = 0.25,
  include.chimeras = TRUE)
}
\arguments{
\item{obj}{An object containing variant counts}

\item{...}{additional arguments}

\item{freq}{minimum frequency snv to return (Default: 0.25)}

\item{include.chimeras}{include chimeric reads when calculating SNV frequencies
(Default: TRUE)}
}
\value{
A vector of SNVs and their frequencies
}
\description{
Find single nucleotide variants (SNVs) above a specified frequency
in a table of variants.
}
\author{
Helen Lindsay
}

