% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/accessors.R
\docType{methods}
\name{getChimeras}
\alias{getChimeras}
\alias{getChimeras,CrisprSet-method}
\title{Get chimeric alignments}
\usage{
getChimeras(obj, ...)

\S4method{getChimeras}{CrisprSet}(obj, ..., sample)
}
\arguments{
\item{obj}{An object containing aligned sequences}

\item{...}{additional arguments}

\item{sample}{The sample name or sample index to return}
}
\value{
A GAlignment object containing the chimeric read groups
}
\description{
Return chimeric alignments from a collection of aligned sequences
}
\examples{
data("gol_clutch1")
chimeras <- getChimeras(gol, sample = 2)
}
\author{
Helen Lindsay
}

