% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInsertionsTable.R
\name{getInsertionsTable}
\alias{getInsertionsTable}
\title{getInsertionsTable}
\usage{
getInsertionsTable(alns, pos = 1L)
}
\arguments{
\item{alns}{A GAlignments object}

\item{pos}{(Integer(1)) The amount by which to shift genomic
coordinates upstream to get coordinates relative to a display region}
}
\value{
A data frame of insertion sequences, genomic and
relative locations
}
\description{
Returns a table of insertion sequences present in
a GAlignments object.  This table is aggregated and used by
plotAlignments.
}
\author{
Helen Lindsay
}
