% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/crispRvariants_data.R
\docType{data}
\name{gol_clutch1}
\alias{gol}
\alias{gol_clutch1}
\title{Variant sequences from golden clutch 1 (Burger et al)}
\format{A CrisprSet object countaining 8 samples}
\usage{
data(gol_clutch1)
}
\value{
A CrisprSet object named "gol"
}
\description{
This dataset is a subset of the crispant data for the golden gene used by Burger et al (submitted).
}
\details{
\itemize{
\item gol The variants as a CrisprSet object
}
}
\keyword{datasets}

