% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alleleLabels.R
\name{indelLabels}
\alias{indelLabels}
\title{indelLabels}
\usage{
indelLabels(alns, rc = FALSE, genome.to.pos = NULL, keep.ops = c("I", "D",
  "N"), regions = NULL, ...)
}
\arguments{
\item{alns}{(GAligments) aligned reads for finding variants}

\item{rc}{Should the variants be displayed with respect to the 
negative strand? (Default: FALSE)}

\item{genome.to.pos}{A vector with names being genomic locations and
values being positions to use in labels (Default: NULL)}

\item{keep.ops}{CIGAR operations to remain in the variant label
(usually indels)}

\item{regions}{IRanges(k) Regions for counting insertions and
deletions.  Insertions on the right border are not counted.}

\item{...}{extra formatting arguments}
}
\value{
A vector of labels for alns
}
\description{
Makes allele labels for insertion / deletion variants
}
\author{
Helen Lindsay
}
