% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{indelPercent}
\alias{indelPercent}
\title{Counts the percentage of reads containing an insertion and/or deletion}
\usage{
indelPercent(alns)
}
\arguments{
\item{alns}{The alignments}
}
\description{
Prints the percentage of target reads that include at least one
insertion or deletion
}
\author{
Helen Lindsay
}

