% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.intersperse}
\alias{.intersperse}
\title{.intersperse}
\usage{
.intersperse(outer, inner)
}
\arguments{
\item{outer}{vector that will be the first and last elements}

\item{inner}{vector that will join elements of outer}
}
\value{
A vector interspersing elements of outer and inner.  If
outer is c(a,b,c) and inner is c(d,e), returns c(a,d,b,e,c)
}
\description{
create a vector of elements in outer interspersed with
elements in inner.  Similar to python zip. No element checking.
}
\examples{
CrispRVariants:::.intersperse(c(1:10), c(1:9)*10)
}
\author{
Helen Lindsay
}
