% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotAlignments.R
\name{makeAlignmentTilePlot}
\alias{makeAlignmentTilePlot}
\title{Makes a basic tile plot of aligned sequences}
\usage{
makeAlignmentTilePlot(m, ref, xlab, plot_text_size, axis_text_size, xtick_labs,
  xtick_breaks, tile_height)
}
\description{
Makes a basic tile plot of aligned sequences
}

