% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/accessors.R
\docType{methods}
\name{mutationEfficiency}
\alias{mutationEfficiency}
\alias{mutationEfficiency,CrisprSet-method}
\title{Get mutation efficiency}
\usage{
mutationEfficiency(obj, ...)

\S4method{mutationEfficiency}{CrisprSet}(obj, ..., snv = c("non_variant",
  "include", "exclude"), include.chimeras = TRUE, exclude.cols = NULL,
  filter.vars = NULL, filter.cols = NULL)
}
\arguments{
\item{obj}{An object containing variant counts}

\item{...}{additional arguments}

\item{snv}{Single nucleotide variants (SNVs) may be considered as mutations ("include"),
treated as ambiguous sequences and not counted at all ("exclude"), or treated as
non-mutations, e.g. sequencing errors or pre-existing SNVs ("non_variant", default)}

\item{include.chimeras}{Should chimeric alignments be counted as variants
when calculating mutation efficiency (Default: TRUE}

\item{exclude.cols}{A vector of names or indices of columns in the variant counts table
that will not be considered when counting mutation efficiency}

\item{filter.vars}{Variants to remove before calculating efficiency.  May be either
a variant size, e.g. "1D", or a particular variant/variant combination, e.g. -5:3D}

\item{filter.cols}{A vector of control sample names.  Any variants present in the control
samples will not be counted towards the efficiency.}
}
\value{
A vector of efficiency statistics per sample and overall
}
\description{
Returns the percentage of sequences that contain at least one mutation.
}
\examples{
data("gol_clutch1")
mutationEfficiency(gol)
}
\author{
Helen Lindsay
}

