% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/initialisers.R
\docType{methods}
\name{narrowAlignments}
\alias{narrowAlignments}
\alias{narrowAlignments,GAlignments,GRanges-method}
\title{Narrow a set of aligned reads to a target region}
\usage{
narrowAlignments(alns, target, ...)

\S4method{narrowAlignments}{GAlignments,GRanges}(alns, target, ...,
  reverse.complement, verbose = FALSE)
}
\arguments{
\item{alns}{A GAlignments object including a metadata column "seq"
containing the sequence}

\item{target}{A GRanges object}

\item{reverse.complement}{Should the aligned reads be reverse complemented?}

\item{verbose}{(Default: FALSE)}
}
\value{
A list of "alignments", the narrowed alignments (GAlignments),
and "genome.ranges", the genomic locations of the cigar operations with
respect to the reference strand.
If reverse.complement is true, "alignments" is not a proper GAlignments object.
The start is displayed with respect to the reference strand, but the cigar
and sequence are displayed with respect to the negative strand.
Use with caution outside of crispRvariants!
}
\description{
Aligned reads are narrowed to the target region.  In
the case of reads with deletions spanning the boundaries of the target,
reads are narrowed to the next aligned base outside of the target
Note that alignments and cigars are reversed if reverse.complement = TRUE
}
\author{
Helen Lindsay
}

