% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotAlignments.R
\docType{methods}
\name{plotAlignments}
\alias{plotAlignments}
\alias{plotAlignments,CrisprSet-method}
\alias{plotAlignments,DNAString-method}
\title{Plot alignments with respect to a reference sequence}
\usage{
plotAlignments(obj, ...)

\S4method{plotAlignments}{CrisprSet}(obj, ..., freq_cutoff = 0,
  top_n = nrow(obj$cigar_freqs), renumbered = obj$pars["renumbered"])

\S4method{plotAlignments}{DNAString}(obj, ..., alns, ins_sites,
  highlight_pam = TRUE, show_plot = FALSE, target_loc = 17,
  pam_start = NA, pam_end = NA, ins_size = 6, legend_cols = 3,
  xlab = NULL, xtick_labs = NULL, xtick_breaks = NULL,
  plot_text_size = 8, axis_text_size = 16, legend_text_size = 16,
  highlight_guide = TRUE, guide_loc = NULL, tile_height = 0.55,
  max_insertion_size = 50)
}
\arguments{
\item{freq_cutoff}{i (integer) only plot variants that occur >= i times
(default: 0, i.e no frequency cutoff)}

\item{top_n}{(integer) Plot only the n most frequent variants
(default: plot all)}

\item{renumbered}{If TRUE, the x-axis is numbered with respect to the target
(default: TRUE)}

\item{alns}{A named character vector of aligned sequences, with insertions removed}

\item{ins_sites}{A table of insertion_sites, which must include cols
named "start", "cigar" and "seq", for the start of the insertion in the
corresponding sequence}

\item{highlight_pam}{should location of PAM with respect to the target site be shown?
(Default: TRUE)  If TRUE, and pam_start and pam_end are not supplied, PAM is inferred
from target_loc}

\item{show_plot}{}

\item{target_loc}{The location of the zero point / cleavage location.  Base n, where
the zero point is between bases n and n+1}

\item{pam_start}{}

\item{pam_end}{}

\item{ins_size}{}

\item{legend_cols}{}

\item{xlab}{}

\item{xtick_labs}{}

\item{xtick_breaks}{}

\item{plot_text_size}{}

\item{axis_text_size}{}

\item{legend_text_size}{}

\item{highlight_guide}{}

\item{guide_loc}{}

\item{tile_height}{}

\item{ref}{The reference sequence}
}
\value{
A ggplot figure
}
\description{
Plot alignments with respect to a reference sequence

Wrapper for CrisprSet$plotVariants.  Optionally filters a
CrisprSet frequency table, then plots variants.  More information in
\code{\link[crispRvariants]{CrisprSet}}

Plots a set of pairwise alignments to a reference sequence.
Alignments should all be the same length as the reference sequences.
This is achieved by removing insertions with respect to the reference,
see \code{\link[crispRvariants]{seqsToAln}} for these alignments.
Insertions are indicated by symbols in the plot and a table showing the
inserted sequences below the plot.
}
\author{
Helen Lindsay
}
\seealso{
\code{\link[crispRvariants]{seqsToAln}}, \code{\link[ggplot2]}
}

