% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotFreqHeatmap.R
\docType{methods}
\name{plotFreqHeatmap}
\alias{plotFreqHeatmap}
\alias{plotFreqHeatmap,CrisprSet-method}
\alias{plotFreqHeatmap,matrix-method}
\title{Plot a table of counts with colours indicating frequency}
\usage{
plotFreqHeatmap(obj, ...)

\S4method{plotFreqHeatmap}{matrix}(obj, ..., col.sums = TRUE,
  row.sums = FALSE, group = NULL, group.colours = NULL,
  as.percent = TRUE, x.axis.title = NULL, x.size = 6, y.size = 8,
  x.angle = 90, legend.text.size = 6, plot.text.size = 2,
  line.width = 1, legend.position = "right",
  legend.key.height = grid::unit(3, "lines"))

\S4method{plotFreqHeatmap}{CrisprSet}(obj, ..., top.n = nrow(obj$cigar_freqs),
  freq.cutoff = 0)
}
\arguments{
\item{obj}{A matrix of counts with rows = feature, columns = sample}

\item{...}{additional arguments}

\item{col.sums}{Include a row of column totals at the top of the
plot (Default: TRUE)}

\item{row.sums}{Include a column of row totals (Default: FALSE)}

\item{group}{Grouping factor for columns.  If supplied, columns are
ordered to match the levels  (Default: NULL)}

\item{group.colours}{Colours for column groups, should match levels of "group".
If "NULL", groups are coloured differently (Default: NULL)}

\item{as.percent}{Should colours represent the percentage of reads
per sample (TRUE) or the actual counts (FALSE)?  (Default: TRUE)}

\item{x.axis.title}{A title for the x-axis.  (Default: NULL)}

\item{x.size}{Font size for x-labels (Default: 16)}

\item{y.size}{Font size for y-labels (Default: 16)}

\item{x.angle}{Angle for x-labels (Default: 90, i.e. vertical)}

\item{legend.text.size}{Font size for legend (Default: 16)}

\item{plot.text.size}{Font size counts within plot (Default: 8)}

\item{line.width}{Line thickness of title box}

\item{legend.position}{The position of the legend (Default: right)}

\item{legend.key.height}{The height of the legend key, as a "unit" object.
(See \code{\link[grid]{unit}}).}

\item{top.n}{Show the n top ranked variants.  Note that if the nth and n+1th
variants have equal rank, they will not be shown.   (Default: show all)}

\item{freq.cutoff}{Show variants with frequency >= freq.cutoff
(Default: 0, i.e. no cutoff)}
}
\value{
The ggplot2 plot of the variant frequencies
}
\description{
Plot a table of counts with colours indicating frequency
}
\examples{
#Load a CrisprSet object for plotting
data("gol_clutch1")

# Plot the frequency heatmap
plotFreqHeatmap(gol)
}

