% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotVariants.R
\docType{methods}
\name{plotVariants}
\alias{plotVariants}
\alias{plotVariants,CrisprSet-method}
\title{Plot alignments, frequencies and location of target sequence}
\usage{
plotVariants(obj, ...)

\S4method{plotVariants}{CrisprSet}(obj, ..., txdb, add.chr = TRUE,
  plotAlignments.args = list(), plotFreqHeatmap.args = list())
}
\arguments{
\item{obj}{}

\item{...}{extra arguments for plot layout}

\item{txdb}{GenomicFeatures:TxDb object}

\item{add.chr}{If target chromosome does not start with "chr", e.g.
"chr5", add the "chr" prefix.  (Default:TRUE)}

\item{plotAlignments.args}{Extra arguments for plotAlignments}

\item{plotFreqHeatmap.args}{Extra arguments for plotFreqHeatmap}
}
\description{
Plot alignments, frequencies and location of target sequence
}

