% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialisers.R
\name{rcAlns}
\alias{rcAlns}
\title{Internal CrispRVariants function for deciding whether to reverse
complement aligned reads}
\usage{
rcAlns(target.strand, reverse.complement)
}
\arguments{
\item{target.strand}{Strand of the target region}

\item{reverse.complement}{Should the alignment be oriented to match the strand}
}
\value{
A logical value indicating whether the narrowed alignment should be reverse complemented.
}
\description{
Guides on the negative strand may be displayed with respect
to either strand.  This function checks whether a guide is on the negative
strand and should be reverse complemented.
}
\author{
Helen Lindsay
}

