% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/initialisers.R
\docType{methods}
\name{readsToTarget}
\alias{alnsToCrisprSet}
\alias{readsToTarget}
\alias{readsToTarget,GAlignments,GRanges-method}
\alias{readsToTarget,character,GRanges-method}
\alias{readsToTargets}
\alias{readsToTargets,character,GRanges-method}
\title{Trims reads to a target region.}
\usage{
readsToTarget(reads, target, ...)

\S4method{readsToTarget}{GAlignments,GRanges}(reads, target, ...,
  reverse.complement = TRUE, collapse.pairs = FALSE, use.consensus = TRUE,
  verbose = FALSE, name = NULL)

\S4method{readsToTarget}{character,GRanges}(reads, target, ..., reference,
  reverse.complement = TRUE, exclude.ranges = GRanges(),
  exclude.names = NA, chimeras = c("ignore", "exclude", "merge"),
  collapse.pairs = FALSE, use.consensus = TRUE, names = NULL,
  verbose = FALSE)

readsToTargets(reads, targets, ...)

\S4method{readsToTargets}{character,GRanges}(reads, targets, ..., references,
  primer.ranges = NULL, reverse.complement = TRUE, collapse.pairs = FALSE,
  use.consensus = TRUE, ignore.strand = TRUE, names = NULL,
  bpparam = BiocParallel::SerialParam(), verbose = FALSE)

alnsToCrisprSet(alns, reference, target, reverse.complement, collapse.pairs,
  names, use.consensus, verbose, ...)
}
\arguments{
\item{reads}{A GAlignments object, or a character vector of the filenames}

\item{target}{A GRanges object specifying the range to narrow alignments to}

\item{...}{Extra arguments for initialising CrisprSet}

\item{reverse.complement}{Should the alignments be oriented to match
the strand of the target? (Default: TRUE)}

\item{collapse.pairs}{If reads are paired, should pairs be collapsed?  (Default: FALSE)}

\item{verbose}{Print progress and statistics (Default: FALSE)}

\item{name}{An experiment name for the reads.  (Default: NULL)}

\item{chimeras}{Flag to determine how chimeric reads are treated.  One of
"ignore", "exclude", and "merge".  Default "ignore", "merge" not implemented yet}

\item{names}{Experiment names for each bam file.  If not supplied, filenames are used.}

\item{primer.ranges}{A set of GRanges, corresponding to the targets.
Read lengths are typically greater than target regions, and it can
be that reads span multiple targets.  If primer.ranges are available,
they can be used to assign such reads to the correct target.}

\item{ignore.strand}{Should strand be considered when finding overlaps?
(See \code{\link[GenomicAlignments]{findOverlaps}} )}

\item{bpparam}{A BiocParallel parameter for parallelising across reads.
Default: no parallelisation.  (See \code{\link[BiocParallel]})}

\item{keep.unpaired}{For paired reads, should unpaired alignments be kept? (Default: TRUE)}

\item{ref}{The reference sequence}

\item{...}{Additional arguments for initialising the CrisprSet
primer.ranges are supplied}
}
\value{
A CrisprRun object

A CrisprSet object
}
\description{
Trims aligned reads to a target region, optionally reverse
complementing the alignments.

Internal crispRvariants function for converting a set of reads to a CrisprSet
}
\author{
Helen Lindsay
}

