% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverseCigar.R
\name{reverseCigar}
\alias{reverseCigar}
\title{Reverses the order of operations in a cigar string}
\usage{
reverseCigar(cigars)
}
\arguments{
\item{cigars}{the cigar strings.}
}
\value{
The reversed cigar string
}
\description{
For example, the string "20M5D15M" would become "15M5D20M"
}

