% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{selectOps}
\alias{selectOps}
\title{selectOps}
\usage{
selectOps(alns, ops = GenomicAlignments::CIGAR_OPS, op.regions = NULL,
  pos = 1L)
}
\arguments{
\item{alns}{(GAlignments)}

\item{ops}{CIGAR operations to consider (Default: all)}

\item{op.regions}{(GRanges) Return operations only in these regions}

\item{pos}{An offset for the cigar ranges}
}
\value{
A GRanges list of opertion locations in reference space
with a metadata column for the operation width in query space.
}
\description{
select CIGAR operations in a region of interest.
}
