% Generated by roxygen2 (4.0.2): do not edit by hand
\name{seqsToAln}
\alias{seqsToAln}
\title{Creates a text alignment from a set of cigar strings}
\usage{
seqsToAln(cigar, dnaseq, del_char = "-", aln_start = NULL,
  target_start = NULL, target_end = NULL)
}
\arguments{
\item{cigar}{A list of cigar strings to align}

\item{dnaseq}{The set of sequences corresponding to the cigars, as DNAStrings}

\item{del_char}{The character to represent deleted bases. Default "-"}

\item{aln_start}{Genomic start locations of aligned sequences. Should be
used in conjunction with target_start and target_end.}

\item{target_start}{Genomic start of the region to be returned.}

\item{target_end}{Genomic end of the region to be returned.}
}
\description{
Creates a one-to-one text alignment of a set of cigar strings with respect
to the reference sequence by collapsing insertions and introducing gaps
across deletions.
}
\details{
When genomic coordinates for the alignment start and the target region
are provided, aligned sequences are cropped to the target region
}
\author{
Helen Lindsay
}

