% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqsToAln.R
\name{seqsToPartialAln}
\alias{seqsToPartialAln}
\title{seqsToPartialAln}
\usage{
seqsToPartialAln(cigars, starts, alns, reference, target, keep,
  xbreaks = NULL, xticks = NULL)
}
\arguments{
\item{cigars}{cigar strings corresponding to alignments}

\item{starts}{Genomic start coordinates of the alignments}

\item{alns}{Character vector of pairwise alignments, with insertions removed}

\item{reference}{Reference sequence}

\item{target}{The complete region spanned by the alignments (GRanges)}

\item{keep}{Region to display (IRanges or GRanges)}

\item{xbreaks}{Mapping between coordinates and plot locations.
If supplied, plot locations are adjusted for gaos (Default: NULL)}

\item{xticks}{Locations for x ticks in plot (Default: NULL)}
}
\description{
Given a character vector of pairwise alignments and a
region to display, trims alignments to the display region
}
\author{
Helen Lindsay
}
