% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotAlignments.R
\name{setDNATileColours}
\alias{setDNATileColours}
\title{Format insertions for plotting}
\usage{
setDNATileColours(m)
}
\arguments{
\item{m}{A matrix with a column named "value" of the characters at each tile position.}
}
\description{
Formats a table of insertions for plotting with \code{\link[crispRvariants]{plotAlignments}}.
}
\author{
Helen Lindsay

Helen Lindsay
}

