% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAlignments.R
\name{setMismatchTileColours}
\alias{setMismatchTileColours}
\title{Sets colours for plotting mismatches in aligned DNA sequences.}
\usage{
setMismatchTileColours(m)
}
\arguments{
\item{m}{A data frame of nucleotides and plotting locations, e.g. created
by \code{\link[CrispRVariants]{transformAlnsToLong}}}
}
\value{
A matrix with additional columns specifying tile and text colours
}
\description{
Sets tile colours for \code{\link[CrispRVariants]{plotAlignments}} with a
DNA alphabet.
}
\author{
Helen Lindsay
}
