% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{splitByPCRPrimers}
\alias{splitByPCRPrimers}
\alias{splitByPCRPrimers,GAlignments,GRanges-method}
\title{Finds overlaps between aligned reads and PCR primers}
\usage{
splitByPCRPrimers(bam, primers, ...)

\S4method{splitByPCRPrimers}{GAlignments,GRanges}(bam, primers, ...,
  tolerance = 0, verbose = TRUE)
}
\arguments{
\item{bam}{A set of aligned reads}

\item{tolerance}{Number of bases by which reads and primers may differ
at each end (Default: 0)}

\item{verbose}{Print number of full and partial matches (Default: TRUE)}

\item{primer_ranges}{A set of ranges that the unclipped reads may map to}

\item{bam}{A GAlignments object}

\item{primer_ranges}{A GRanges object}
}
\value{
A \code{\link[IRanges]{Hits}} object where "query" is the index with
respect to bam and "subject" is the index with respect to the primers.
}
\description{
Short reads amplified with PCR primers should start and end
at defined positions.  However, the ends of an aligned read may be clipped
as sequencing technologies are prone to making errors at the start and end.
splitByPCRPrimers extrapolates the genomic location of entire reads
from their aligned sections by adding clipped sections, then finds near
exact matches to a set of PCR primers.
}
\author{
Helen Lindsay
}
\seealso{
\code{\link[GenomicRanges]{GRanges}}, \code{\link[GenomicAlignments]{GAlignments}}
}

