% Generated by roxygen2 (4.0.2): do not edit by hand
\name{abifToFastq}
\alias{abifToFastq}
\title{Read a file in ab1 (Sanger) format and convert to fastq}
\usage{
abifToFastq(seqname, fname, outfname, trim = TRUE, cutoff = 0.05,
  min_seq_len = 20, offset = 33)
}
\arguments{
\item{seqname}{name of sequence, to appear in fastq file}

\item{fname}{filename of sequence in ab1 format}

\item{outfname}{filename to append the fastq output to}

\item{trim}{should low quality bases be trimmed from the ends?  TRUE or FALSE}

\item{cutoff}{probability cutoff}

\item{min_seq_len}{minimum number of sequenced bases required in order to trim the read}

\item{offset}{phred offset for quality scores}
}
\description{
This is an R implementation of Wibowo Arindrarto's abifpy.py trimming module,
which itself implement's Richard Mott's trimming algorithm
See \link{https://github.com/bow/abifpy} for more details.
}
\details{
Requires Bioconductor package SangerseqR
}
\author{
Helen Lindsay
}

