% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readsByPCRPrimer.R
\docType{methods}
\name{addClipped}
\alias{addClipped}
\alias{addClipped,GAlignments-method}
\title{Extrapolates mapping location from clipped, aligned reads}
\usage{
addClipped(bam, ...)

\S4method{addClipped}{GAlignments}(bam, ...)
}
\arguments{
\item{bam}{A GAlignments object}

\item{...}{additional arguments}
}
\value{
A \code{\link[GenomicRanges]{GRanges}} representation of the extended
mapping locations
}
\description{
Extrapolates the mapping location of a read by assuming that
the clipped regions should map adjacent to the mapped locations.  This
is not always a good assumption, particularly in the case of chimeric reads!
}
\author{
Helen Lindsay
}

