% Generated by roxygen2 (4.0.2): do not edit by hand
\name{barplotAlleleFreqs}
\alias{barplotAlleleFreqs}
\title{Plots barplots of the spectrum of variants for a sample set}
\usage{
barplotAlleleFreqs(allele_counts, group = NULL, bar_colours = NULL,
  group_colours = NULL, legend_text_size = 10, legend_symbol_size = 1,
  show_percentage = TRUE, snv_label = "SNV", novar_label = "no variant")
}
\arguments{
\item{allele_counts}{}

\item{group}{}

\item{bar_colours}{}

\item{group_colours}{}

\item{legend_text_size}{}

\item{legend_symbol_size}{}

\item{show_percentage}{}

\item{snv_label}{}

\item{novar_label}{}
}
\description{
Given a matrix of counts with columns being samples and rows being
variant alleles, groups variants by size and type, and produces a
barplot showing the variant spectrum for each sample.
}
\author{
Helen Lindsay
}

