% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/excludeFromBam.R
\name{excludeFromBam}
\alias{excludeFromBam}
\title{Removes reads from a bam file}
\usage{
excludeFromBam(bam, exclude.ranges = GRanges(), exclude.names = NA)
}
\arguments{
\item{bam}{a GAlignments object}

\item{exclude.ranges}{Regions to exclude, as \code{\link[GenomicRanges]{GRanges}}.}

\item{exclude.names}{A character vector of alignments names to exclude}
}
\value{
The bam minus the excluded regions
}
\description{
Returns a GAlignments excluding reads based on either name and/or location
}
\author{
Helen Lindsay
}

