% Generated by roxygen2 (4.0.2): do not edit by hand
\name{excludeFromBamByName}
\alias{excludeFromBamByName}
\title{Removes reads from a bam file}
\usage{
excludeFromBamByName(bam, exclude_names)
}
\arguments{
\item{bam}{A GAlignments object}

\item{exclude_names}{a character vector of sequence names to exclude}
}
\description{
Returns a GAlignments excluding reads based on name
}
\author{
Helen Lindsay
}

