% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/excludeFromBam.R
\name{excludeFromBamByRange}
\alias{excludeFromBamByRange}
\title{Removes reads from a bam file}
\usage{
excludeFromBamByRange(bam, exclude.ranges)
}
\arguments{
\item{bam}{a GAlignments object}

\item{exclude.ranges}{A GRanges object containing regions to exclude}
}
\description{
Returns a GAlignments excluding reads based on location
}
\author{
Helen Lindsay
}

