% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/findChimeras.R
\name{findChimeras}
\alias{findChimeras}
\title{Find chimeric reads}
\usage{
findChimeras(bam)
}
\arguments{
\item{bam}{A GAlignments object, must include names}
}
\description{
Assumes the the GAlignments object does not contain multimapping reads.
That is, read names that appear more than ones in the file are considered
chimeras.
}
\author{
Helen Lindsay
}

