% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAlignments.R
\name{makeAlignmentTilePlot}
\alias{makeAlignmentTilePlot}
\title{Internal CrispRVariants function for creating the plotAlignments background}
\usage{
makeAlignmentTilePlot(m, ref, xlab, plot.text.size, axis.text.size, xtick.labs,
  xtick.breaks, tile.height)
}
\arguments{
\item{m}{A matrix with column headings Var1: y location, Var2: x location, 
cols: tile fill colour, isref: transparency value text_cols: text colour}

\item{ref}{The reference sequence, only used for checking the number of
x-tick labels when x-tick breaks are not supplied}

\item{xlab}{Label for the x axis}

\item{plot.text.size}{Size for text within plot}

\item{axis.text.size}{Size for text on axes}

\item{xtick.labs}{x axis labels}

\item{xtick.breaks}{Locations of x labels}

\item{tile.height}{Controls whitespace between tiles}
}
\value{
A ggplot object
}
\description{
Takes a matrix of characters, x and y locations and colours, creates
a ggplot geom_tile plot with tiles labelled by the characters.
}
\author{
Helen Lindsay
}
