% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotAlignments.R
\name{makeAlignmentTilePlot}
\alias{makeAlignmentTilePlot}
\title{Makes a basic tile plot of aligned sequences}
\usage{
makeAlignmentTilePlot(m, ref, xlab, plot.text.size, axis.text.size, xtick.labs,
  xtick.breaks, tile.height)
}
\description{
Makes a basic tile plot of aligned sequences
}

