% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/accessors.R
\docType{methods}
\name{mutationEfficiency}
\alias{mutationEfficiency}
\alias{mutationEfficiency,CrisprSet-method}
\title{Get mutation efficiency}
\usage{
mutationEfficiency(obj, ...)

\S4method{mutationEfficiency}{CrisprSet}(obj, ..., snv = c("include",
  "exclude", "non_variant"), exclude.cols = NULL)
}
\arguments{
\item{obj}{An object containing variant counts}

\item{...}{additional arguments}

\item{snv}{Single nucleotide variants (SNVs) may be considered as mutations ("include"),
treated as ambiguous sequences and not counted at all ("exclude"), or treated as
non-mutations, e.g. sequencing errors or pre-existing SNVs ("non_variant")}

\item{exclude.cols}{A vector of names or indices of columns in the variant counts table
that will not be considered when counting mutation efficiency}
}
\value{
A vector of efficiency statistics per sample and overall
}
\description{
Returns the percentage of sequences that contain at least one mutation.
}
\examples{
data("gol_clutch1")
mutationEfficiency(gol)
}
\author{
Helen Lindsay
}

