% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotAlignments.R
\docType{methods}
\name{plotAlignments}
\alias{plotAlignments}
\alias{plotAlignments,CrisprSet-method}
\alias{plotAlignments,DNAString-method}
\title{Plot alignments with respect to a reference sequence}
\usage{
plotAlignments(obj, ...)

\S4method{plotAlignments}{CrisprSet}(obj, ..., freq.cutoff = 0,
  top.n = nrow(obj$cigar_freqs), renumbered = obj$pars["renumbered"])

\S4method{plotAlignments}{DNAString}(obj, ..., alns, ins.sites,
  highlight.pam = TRUE, show.plot = FALSE, target.loc = 17,
  pam.start = NA, pam.end = NA, ins.size = 6, legend.cols = 3,
  xlab = NULL, xtick.labs = NULL, xtick.breaks = NULL,
  plot.text.size = 8, axis.text.size = 16, legend.text.size = 16,
  highlight.guide = TRUE, guide.loc = NULL, tile.height = 0.55,
  max.insertion.size = 50)
}
\arguments{
\item{freq.cutoff}{i (integer) only plot variants that occur >= i times
(default: 0, i.e no frequency cutoff)}

\item{top.n}{(integer) Plot only the n most frequent variants
(default: plot all)}

\item{renumbered}{If TRUE, the x-axis is numbered with respect to the target
(default: TRUE)}

\item{alns}{A named character vector of aligned sequences, with insertions removed}

\item{ins.sites}{A table of insertion_sites, which must include cols
named "start", "cigar" and "seq", for the start of the insertion in the
corresponding sequence}

\item{highlight.pam}{should location of PAM with respect to the target site be shown?
(Default: TRUE)  If TRUE, and pam.start and pam.end are not supplied, PAM is inferred
from target.loc}

\item{show.plot}{}

\item{target.loc}{The location of the zero point / cleavage location.  Base n, where
the zero point is between bases n and n+1}

\item{pam.start}{}

\item{pam.end}{}

\item{ins.size}{}

\item{legend.cols}{}

\item{xlab}{}

\item{xtick.labs}{}

\item{xtick.breaks}{}

\item{plot.text.size}{}

\item{axis.text.size}{}

\item{legend.text.size}{}

\item{highlight.guide}{}

\item{guide.loc}{}

\item{tile.height}{}

\item{ref}{The reference sequence}
}
\value{
A ggplot figure
}
\description{
Plot alignments with respect to a reference sequence

Wrapper for CrisprSet$plotVariants.  Optionally filters a
CrisprSet frequency table, then plots variants.  More information in
\code{\link[crispRvariants]{CrisprSet}}

Plots a set of pairwise alignments to a reference sequence.
Alignments should all be the same length as the reference sequences.
This is achieved by removing insertions with respect to the reference,
see \code{\link[crispRvariants]{seqsToAln}} for these alignments.
Insertions are indicated by symbols in the plot and a table showing the
inserted sequences below the plot.
}
\author{
Helen Lindsay
}
\seealso{
\code{\link[crispRvariants]{seqsToAln}}, \code{\link[ggplot2]}
}

