% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotChimeras}
\alias{plotChimeras}
\title{Display a dot plot of chimeric alignments}
\usage{
plotChimeras(chimeric_alns, max_gap = 10, tick_sep = 20, text_size = 12,
  title_size = 16, legend_title = "Chromosome")
}
\arguments{
\item{chimeric_alns}{A GAlignments object containing only the chimeric
reads to be plotted}

\item{max_gap}{The maximum gap between aligned blocks allowed
aligned before blocks are collapsed, in bases.  Default 10}

\item{tick_sep}{How many bases should separate tick labels on plot.  Default 20.}

\item{text_size}{Size of X and Y tick labels on plot.  Default 12}

\item{title_size}{Size of X and Y axis labels on plot.  Default 16}

\item{legend_title}{Title for the legend.  Default "Chromosome"}
}
\description{
Produces a dot plot of a set of chimeric alignments.  For chimeric alignments,
a single read is split into several, possibly overlapping alignmed blocks.
Aligned sections of chimeric reads can be separated by large genomic distances,
or on separate chromosomes.  plotChimeras produces a dot plot, each aligned block
highlighted, and chromosomes shown in different colours. Large gaps between
aligned segments are collapsed and indicated on the plot with horizontal lines.
The X-axis shows each base of the entire read.
}
\author{
Helen Lindsay
}

