% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialisers.R
\name{rcAlns}
\alias{rcAlns}
\title{Internal CrispRVariants function for determining read orientation}
\usage{
rcAlns(target.strand, orientation)
}
\arguments{
\item{target.strand}{The target strand (one of "+","-","*")}

\item{orientation}{One of "target", "opposite" and "positive" (Default: "target")}
}
\value{
A logical value indicating whether reads should be reverse complemented
}
\description{
Function for determining whether reads should be oriented to the
target strand, always displayed on the positive strand, or oriented to
}
\author{
Helen Lindsay
}
