% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/initialisers.R
\name{rcAlns}
\alias{rcAlns}
\title{Internal crispRvariants function for deciding whether to reverse
complement aligned reads}
\usage{
rcAlns(target.strand, reverse.complement)
}
\arguments{
\item{target.strand}{Strand of the target region}

\item{reverse.complement}{Should the alignment be oriented to match the strand}
}
\value{
A logical value indicating whether the narrowed alignment should be reverse complemented.
}
\description{
Internal crispRvariants function for deciding whether to reverse
complement aligned reads
}
\author{
Helen Lindsay
}

