% Generated by roxygen2 (4.0.2): do not edit by hand
\name{reverseCigar}
\alias{reverseCigar}
\title{Reverses the order of operations in a cigar string}
\usage{
reverseCigar(cigar)
}
\arguments{
\item{cigar}{the cigar string.}
}
\description{
For example, the string "20M5D15M" would become "15M5D20M"
}

