% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/readsByPCRPrimer.R
\docType{methods}
\name{rmMultiPCRChimera}
\alias{rmMultiPCRChimera}
\alias{rmMultiPCRChimera,character,Hits,integer-method}
\title{Remove chimeric reads overlapping multiple primers}
\usage{
rmMultiPCRChimera(readnames, pcrhits, chimera_idxs, ...)

\S4method{rmMultiPCRChimera}{character,Hits,integer}(readnames, pcrhits,
  chimera_idxs, ..., verbose = TRUE)
}
\arguments{
\item{readnames}{A set of read names, used for identifying chimeric read sets}

\item{pcrhits}{A mapping between indices of reads and a set of pcr primers}

\item{chimera_idxs}{location of chimeric reads within the bam}

\item{verbose}{}

\item{pcrhits}{A hits object where queryHits are indices in the bam and subjectHits are
indices within a list of pcr primer ranges.}
}
\value{
pcrhits, with chimeric reads mapping to different primers omitted.
}
\description{
Remove chimeric reads overlapping multiple primers
}
\author{
Helen Lindsay
}

