% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{methods}
\name{selectOps}
\alias{selectOps}
\alias{selectOps,character-method}
\title{selectOps}
\usage{
selectOps(cigar, ...)

\S4method{selectOps}{character}(cigar, ...,
  ops = GenomicAlignments::CIGAR_OPS, op.regions = NULL, pos = 1L)
}
\arguments{
\item{cigar}{CIGAR strings}

\item{...}{Extra arguments (Not currently used)}

\item{ops}{CIGAR operations to consider (Default: all)}

\item{op.regions}{(GRanges) Return operations only in these regions}

\item{pos}{An offset for the cigar ranges}
}
\value{
A GRanges list of opertion locations in reference space
with a metadata column for the operation width in query space.
}
\description{
select CIGAR operations in a region of interest.
}
\author{
Helen Lindsay
}
