% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqsToAln.R
\name{seqsToAln}
\alias{seqsToAln}
\title{Creates a text alignment from a set of cigar strings}
\usage{
seqsToAln(cigar, dnaseq, target, del_char = "-", aln_start = NULL)
}
\arguments{
\item{cigar}{A list of cigar strings to align}

\item{dnaseq}{The set of sequences corresponding to the cigars, as Biostrings::DNAStrings}

\item{target}{The target region to return, as GRanges.  Sequences overlapping
the target region are trimmed to exactly match it.}

\item{del_char}{The character to represent deleted bases. Default "-"}

\item{aln_start}{Genomic start locations of aligned sequences. Should be
used in conjunction with target_start and target_end.}
}
\value{
The sequences with insertions collapsed and deletions padded
}
\description{
Creates a one-to-one text alignment of a set of cigar strings with respect
to the reference sequence by collapsing insertions and introducing gaps
across deletions.

When genomic coordinates for the alignment start and the target region
are provided, aligned sequences are cropped to the target region
}
\author{
Helen Lindsay
}

