% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAlignments.R
\name{setDNATileColours}
\alias{setDNATileColours}
\title{Sets colours for plotting aligned DNA sequences.}
\usage{
setDNATileColours(m)
}
\arguments{
\item{m}{A matrix with a column named "value" of the characters at each tile position.}
}
\value{
A matrix with additional columns specifying tile and text colours
}
\description{
Sets tile colours for \code{\link[CrispRVariants]{plotAlignments}} with a
DNA alphabet.  Colour names must be valid.
}
\author{
Helen Lindsay
}

