% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotAlignments.R
\name{transformAlnsToLong}
\alias{transformAlnsToLong}
\title{Transform data for plotting}
\usage{
transformAlnsToLong(ref, alns)
}
\arguments{
\item{ref}{}

\item{alns}{}
}
\description{
Orders and transforms a reference sequence and a set of aligned sequences
into long format, i.e. one observation (tile position) per row.  Used internally by
\code{\link[crispRvariants]{plotAlignments}}.
}
\author{
Helen Lindsay
}

