% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotAlignments.R
\name{transformAlnsToLong}
\alias{transformAlnsToLong}
\title{Transform data for plotting}
\usage{
transformAlnsToLong(ref, alns, add.other = FALSE)
}
\arguments{
\item{ref}{The reference sequence}

\item{alns}{Character vector of aligned sequences}

\item{add.other}{Add a blank row labelled "Other" (Default: FALSE)}
}
\value{
A matrix of characters and plotting locations
}
\description{
Orders and transforms a reference sequence and a set of aligned sequences
into long format, i.e. one observation (tile position) per row.  Used internally by
\code{\link[crispRvariants]{plotAlignments}}.
}
\author{
Helen Lindsay
}

