% Generated by roxygen2 (4.0.2): do not edit by hand
\name{writeFastq}
\alias{writeFastq}
\title{Append a sequence to a fastq file}
\usage{
writeFastq(outf, vals, allow_spaces = FALSE)
}
\arguments{
\item{outf}{Name of fastq file to append sequence to}

\item{vals}{A list containing entries named "seq" (sequence) and
"quals" (quality scores, in ASCII format)}

\item{allow_spaces}{Should spaces in the sequence name be
substituted with underscores?  TRUE or FALSE}
}
\description{
Used by abifToFastq
}
\author{
Helen Lindsay
}

