% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/writeFastq.R
\name{writeFastq}
\alias{writeFastq}
\title{Append a sequence to a fastq file

Used by abifToFastq}
\usage{
writeFastq(outf, vals, allow_spaces = FALSE, check = TRUE)
}
\arguments{
\item{outf}{Name of fastq file to append sequence}

\item{vals}{A list containing entries named "seq" (sequence) and
"quals" (quality scores, in ASCII format)}

\item{allow_spaces}{Should spaces in the sequence name be
substituted with underscores?  TRUE or FALSE}

\item{check}{Check whether reads with the same name already
exist in the output fastq.  (Default: TRUE)}
}
\value{
None.  The sequences in "vals" are written to outf
}
\description{
Append a sequence to a fastq file

Used by abifToFastq
}
\author{
Helen Lindsay
}

